feat: create AppCheckboxModal and apply to onboarding screen
