+++ App.tsx
@@ -3,7 +3,7 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { enableScreens } from 'react-native-screens';
 import RootNavigator from './src/navigation/RootNavigator';
-import SplashScreen from './src/screens/onboarding/SplashScreen';
+import SplashScreen from './src/screens/splash/SplashScreen';
 
 enableScreens();
