+++ src/screens/onboarding/OnboardingScreen.tsx
@@ -6,20 +6,21 @@ import {
   Image,
   Modal,
   SafeAreaView,
-  StyleSheet,
   View,
 } from 'react-native';
 import { AppText } from '../../components/AppText';
 import { AppButton } from '../../components/AppButton';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { RootStackParamList } from '../../navigation/RootStackParamList';
-import { colors } from '../../theme/color';
 import { TouchableOpacity } from 'react-native';
-import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
-import { AppModal } from '../../components/AppModal';
+import { AppModal } from '../../components/modal/AppModal';
+import { colors } from '../../theme/color';
+import { styles } from './styles';
+import { AppCheckbox } from '../../components/AppCheckbox';
+import { AppCheckboxModal } from '../../components/modal/AppCheckboxModal';
 
 // 화면 너비, 높이 가져오기 (페이지 단위 스크롤)
-const { width, height } = Dimensions.get('window');
+const { width } = Dimensions.get('window');
 
 // 타이틀 데이터 배열 정의
 const TitleData = [
@@ -60,50 +61,6 @@ export default function OnboardingScreen() {
     setAgreePrivacy(next);
   };
 
-  // AgreementItem
-  const AgreementItem = ({
-    checked,
-    onPress,
-    label,
-    bold,
-    highlight,
-  }: {
-    checked: boolean;
-    onPress: () => void;
-    label: React.ReactNode;
-    bold?: boolean;
-    highlight?: boolean;
-  }) => (
-    <TouchableOpacity onPress={onPress}>
-      <View style={styles.agreementItem}>
-        <View style={[styles.checkboxBase, checked && styles.checkboxChecked]}>
-          {checked && <MaterialIcons name="check" size={16} color="white" />}
-        </View>
-        <AppText
-          style={[
-            styles.label,
-            bold && styles.boldLabel,
-            {
-              color: highlight
-                ? checked
-                  ? colors.primary1
-                  : colors.gray3
-                : colors.black,
-              textDecorationLine: highlight && checked ? 'underline' : 'none',
-            },
-          ]}
-        >
-          {label}
-        </AppText>
-      </View>
-    </TouchableOpacity>
-  );
-
-  // 네비게이션 객체 사용
-  type OnboardingScreenNavigationProp = NativeStackNavigationProp<
-    RootStackParamList,
-    'Onboarding'
-  >;
   const navigation =
     useNavigation<NativeStackNavigationProp<RootStackParamList>>();
 
@@ -114,7 +71,6 @@ export default function OnboardingScreen() {
   const [currentIndex, setCurrentIndex] = useState(0);
 
   // 페이지 인디케이터 설정
-  // eslint-disable-next-line react/no-unstable-nested-components
   const PageIndicator = ({
     count,
     currentIndex,
@@ -172,17 +128,12 @@ export default function OnboardingScreen() {
         horizontal
         pagingEnabled
         showsHorizontalScrollIndicator={false}
-        // onMomentumScrollEnd: 스크롤이 끝났을 때 호출되는 이벤트
         onMomentumScrollEnd={e => {
-          // e.nativeEvent.contentOffset.x : 수평 스크롤 거리 (px)
-          // Math.round(... / width) : 현재 몇 번째 인덱스인지 계산
           const newIndex = Math.round(e.nativeEvent.contentOffset.x / width);
           setCurrentIndex(newIndex);
         }}
       />
-
       <PageIndicator count={TitleData.length} currentIndex={currentIndex} />
-
       <View style={styles.bottomButtonContainer}>
         {currentIndex === TitleData.length - 1 ? (
           <View style={{ gap: 16 }}>
@@ -218,229 +169,20 @@ export default function OnboardingScreen() {
         )}
       </View>
 
-      {/* 약간 동의 모달 */}
-      <Modal
-        animationType="slide"
-        transparent={true}
+      {/* 이용 약관 모달 */}
+      <AppCheckboxModal
         visible={modalVisible}
-        onRequestClose={() => setModalVisible(false)}
-      >
-        <View style={styles.modalContainer}>
-          <View style={styles.modalContent}>
-            {/* 타이틀 */}
-            <View
-              style={{ flexDirection: 'row', justifyContent: 'space-between' }}
-            >
-              <AppText color="black" size="lg" weight="bold">
-                서비스 이용 약관에 동의해주세요.
-              </AppText>
-              <TouchableOpacity onPress={() => setModalVisible(false)}>
-                <Image
-                  source={require('../../assets/images/icons/cancel.png')}
-                  style={{ width: 24, height: 24 }}
-                />
-              </TouchableOpacity>
-            </View>
-
-            {/* 약관 체크박스 */}
-            <View style={{ marginTop: 16, marginBottom: 32 }}>
-              <AgreementItem
-                checked={allChecked}
-                onPress={toggleAll}
-                label="모두 동의"
-              />
-              <View style={styles.divider} />
-              <AgreementItem
-                checked={agree14}
-                onPress={() => setAgree14(!agree14)}
-                label={
-                  <AppText
-                    size="sm"
-                    weight="medium"
-                    style={{
-                      color: agree14 ? colors.gray4 : colors.gray3,
-                    }}
-                  >
-                    (필수) 만 14세 이상입니다
-                  </AppText>
-                }
-              />
-              <View style={{ flexDirection: 'row' }}>
-                <AgreementItem
-                  checked={agreeService}
-                  onPress={() => setAgreeService(!agreeService)}
-                  label={
-                    <>
-                      <AppText
-                        size="sm"
-                        weight="medium"
-                        style={{
-                          color: agreeService ? colors.gray4 : colors.gray3,
-                        }}
-                      >
-                        (필수){' '}
-                      </AppText>
-                      <AppText
-                        size="sm"
-                        weight="medium"
-                        style={{
-                          color: agreeService ? colors.primary1 : colors.gray3,
-                        }}
-                        onPress={() => setAgreeServiceModal(true)}
-                      >
-                        서비스 이용약관 확인
-                      </AppText>
-                    </>
-                  }
-                  highlight
-                />
-              </View>
-              <View style={{ flexDirection: 'row' }}>
-                <AgreementItem
-                  checked={agreePrivacy}
-                  onPress={() => setAgreePrivacy(!agreePrivacy)}
-                  label={
-                    <>
-                      <AppText
-                        size="sm"
-                        weight="medium"
-                        style={{
-                          color: agreePrivacy ? colors.gray4 : colors.gray3,
-                        }}
-                      >
-                        (필수){' '}
-                      </AppText>
-                      <AppText
-                        size="sm"
-                        weight="medium"
-                        style={{
-                          color: agreePrivacy ? colors.primary1 : colors.gray3,
-                        }}
-                        onPress={() => setAgreePrivacyModal(true)}
-                      >
-                        개인정보 수집이용 동의
-                      </AppText>
-                    </>
-                  }
-                  highlight
-                />
-              </View>
-
-              {/* 동의 약관 내용 모달들 */}
-              <AppModal
-                title={'서비스 이용약관'}
-                content={'대충 서비스 이용약관 내용입니다.'}
-                visible={agreeServiceModal}
-                onConfirm={() => setAgreeServiceModal(false)}
-                onCancel={() => setAgreeServiceModal(false)}
-              />
-              <AppModal
-                title={'개인정보 수집이용'}
-                content={'대충 개인정보 수집이용 내용입니다.'}
-                visible={agreePrivacyModal}
-                onConfirm={() => setAgreePrivacyModal(false)}
-                onCancel={() => setAgreePrivacyModal(false)}
-              />
-            </View>
-
-            {/* 하단 버튼 */}
-            <AppButton
-              activate={agree14 && agreeService && agreePrivacy}
-              title="동의하고 계속하기"
-              onPress={() => {
-                setModalVisible(false);
-                navigation.navigate('Signup');
-              }}
-            />
-          </View>
-        </View>
-      </Modal>
+        title="서비스 이용 약관에  동의해주세요."
+        firstContent="(필수) 만 14세 이상입니다"
+        checkLabel="(필수) "
+        secondContent="서비스 이용약관 확인"
+        thirdContent="개인정보 수집이용 동의"
+        onClose={() => setModalVisible(false)}
+        onAgree={() => {
+          setModalVisible(false);
+          navigation.navigate('Signup');
+        }}
+      />
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  page: {
-    width,
-    alignItems: 'center',
-    justifyContent: 'flex-start',
-    paddingTop: height * 0.12,
-  },
-  modalContainer: {
-    height: '100%',
-    backgroundColor: 'rgba(0,0,0,0.5)',
-    justifyContent: 'flex-end',
-  },
-  modalContent: {
-    backgroundColor: 'white',
-    padding: 24,
-    borderTopLeftRadius: 16,
-    borderTopRightRadius: 16,
-  },
-  image: {
-    width: width * 0.9,
-    height: height * 0.3,
-    marginVertical: height * 0.1,
-  },
-  bottomButtonContainer: {
-    position: 'absolute',
-    bottom: height * 0.1,
-    left: 20,
-    right: 20,
-  },
-  indicatorContainer: {
-    position: 'absolute',
-    bottom: height * 0.3,
-    width: '100%',
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  dot: {
-    width: 6,
-    height: 6,
-    borderRadius: 3,
-    marginHorizontal: 4,
-  },
-  activeDot: {
-    backgroundColor: colors.primary1,
-  },
-  inactiveDot: {
-    backgroundColor: colors.gray1,
-  },
-  agreementItem: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginVertical: 8,
-  },
-  label: {
-    fontSize: 14,
-    marginLeft: 8,
-    color: colors.black,
-  },
-  boldLabel: {
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-  highlightLabel: {
-    color: colors.primary1,
-  },
-  checkboxBase: {
-    width: 16,
-    height: 16,
-    borderWidth: 2,
-    borderColor: colors.gray3,
-    borderRadius: 2,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  checkboxChecked: {
-    backgroundColor: colors.primary1,
-    borderWidth: 0,
-  },
-  divider: {
-    height: 1,
-    backgroundColor: colors.gray1,
-    marginVertical: 12,
-  },
-});
